Actualización de tabla de convocatorias automaticas
Al registrar una convocatoria se actualiza la tabla que muestra las convocatorias automaticamente
