+++ src/components/captacion/RegistroConvocatoria.jsx
@@ -11,7 +11,7 @@ import * as Yup from "yup";
 
 const apiUrl = import.meta.env.VITE_API_URL;
 
-const RegistroConvocatoria = () => {
+const RegistroConvocatoria = ({ onRegistroExitoso }) => {
   const toast = useRef(null);
 
   const lugares = [
@@ -84,45 +84,41 @@ const RegistroConvocatoria = () => {
         .required("El número máximo de participantes es obligatorio"),
     }),
 
-    onSubmit: async (values) => {
-      try {
-        let token = JSON.parse(localStorage.getItem("access-token"));
-    
-        if (!token) {
-          token = await refreshToken();
-        }
-    
-        // Formatear las fechas al formato 'YYYY-MM-DD'
-        const fechaRegistroFormateada = new Date(values.fechaRegistro).toISOString().split("T")[0];
-        const fechaEntregaFormateada = new Date(values.fechaEntregaResultados).toISOString().split("T")[0];
-    
-        // Preparar los datos de la solicitud
-        const requestData = {
-          lugar_convocatoria: values.lugarConvocatoria, 
-          fecha_limite_registro: fechaRegistroFormateada,
-          fecha_entrega_resultados: fechaEntregaFormateada,
-          max_participantes: parseInt(values.maxParticipantes, 10), 
-        };
-    
-    
-        // Realizar la solicitud POST
-        const response = await axios.post(
-          `${apiUrl}/captacion/convocatorias/`,
-          requestData,
-          {
-            headers: {
-              Authorization: `Bearer ${token}`,
-              "Content-Type": "application/json",
-            },
-          }
-        );
-    
-        toast.current.show({
-          severity: "success",
-          summary: "Convocatoria registrada",
-          detail: "La convocatoria se registró exitosamente",
-          life: 3000,
-        });
+        onSubmit: async (values) => {
+            try {
+                let token = JSON.parse(localStorage.getItem("access-token"));
+                if (!token) {
+                    token = await refreshToken();
+                }
+
+                const fechaRegistroFormateada = new Date(values.fechaRegistro).toISOString().split("T")[0];
+                const fechaEntregaFormateada = new Date(values.fechaEntregaResultados).toISOString().split("T")[0];
+
+                const requestData = {
+                    lugar_convocatoria: values.lugarConvocatoria,
+                    fecha_limite_registro: fechaRegistroFormateada,
+                    fecha_entrega_resultados: fechaEntregaFormateada,
+                    max_participantes: parseInt(values.maxParticipantes, 10),
+                };
+
+                await axios.post(`${apiUrl}/captacion/convocatorias/`, requestData, {
+                    headers: {
+                        Authorization: `Bearer ${token}`,
+                        "Content-Type": "application/json",
+                    },
+                });
+
+                toast.current.show({
+                    severity: "success",
+                    summary: "Convocatoria registrada",
+                    detail: "La convocatoria se registró exitosamente",
+                    life: 3000,
+                });
+
+                // Llamamos a la función de actualización
+                if (onRegistroExitoso) {
+                    onRegistroExitoso();
+                }
       } catch (error) {
         const errorMessage =
           error.response?.data?.detail || "Hubo un problema al registrar la convocatoria";
